requirements.txt
discord.py>=2.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0
aiohttp>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
schedule>=1.2.0 